backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
groq
python-dotenv
